auth, file upload

+++ app/Http/Controllers/PostsController.php
@@ -5,9 +5,20 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Post;
 use DB;
+use Illuminate\Support\Facades\Storage;
 
 class PostsController extends Controller
 {
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -44,13 +55,32 @@ class PostsController extends Controller
     {
         $this->validate($request, [
                 'title' => 'required',
-                'body'  => 'required'
+                'body'  => 'required',
+                'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        // handle file upload
+        if($request->hasFile('cover_image')){
+            // get filename with extension
+            $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // get just filename
+            $filename = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            // get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // filename to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            // upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }else{
+            $fileNameToStore = 'noimage.jpg';
+        }
 
         // Create Post
         $post = new Post;
         $post->title = $request->input('title');
         $post->body  = $request->input('body');
+        $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
         return redirect('posts')->with('success', 'Post Created');
     }
@@ -76,6 +106,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        
+        // check for correct user
+        if(auth()->user()->id != $post->user->id)
+            return redirect('/posts')->with('error', 'Unauthorized page');
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -90,13 +125,30 @@ class PostsController extends Controller
     {
         $this->validate($request, [
                 'title' => 'required',
-                'body'  => 'required'
+                'body'  => 'required',
+                'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        // handle file upload
+        if($request->hasFile('cover_image')){
+            // get filename with extension
+            $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // get just filename
+            $filename = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            // get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // filename to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            // upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }
 
         // Create Post
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body  = $request->input('body');
+        if($request->hasFile('cover_image'))
+            $post->cover_image = $fileNameToStore;
         $post->save();
         return redirect('posts')->with('success', 'Post Updated');
     }
@@ -110,6 +162,14 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        // check for correct user
+        if(auth()->user()->id != $post->user->id)
+            return redirect('/posts')->with('error', 'Unauthorized page');
+        if($post->cover_image != 'noimage.jpg'){
+            // delete image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
+
         $post->delete();
         return redirect('posts')->with('success', 'Post Removed');
     }
